+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/AbstractApexUnitTestRule.java
@@ -32,7 +32,7 @@ public abstract class AbstractApexUnitTestRule extends AbstractApexRule {
     }
 
     protected boolean isTestMethodOrClass(final ApexNode<?> node) {
-        final ASTModifierNode modifierNode = node.getFirstChildOfType(ASTModifierNode.class);
+        final ASTModifierNode modifierNode = node.firstChild(ASTModifierNode.class);
         return modifierNode != null && modifierNode.isTest();
     }
 }